#4003: serialize tensor with memory config. Display memory configs in the tensor report

+++ tests/ttnn/conftest.py
@@ -29,6 +29,7 @@ def pre_and_post():
     ttnn.load_config_from_json_file(ttnn.CONFIG_PATH)
     ttnn.load_config_from_dictionary(json.loads(ttnn.CONFIG_OVERRIDES))
     logger.debug(f"ttnn.CONFIG:\n{pprint.pformat(dataclasses.asdict(ttnn.CONFIG))}")
-    ttnn.database.delete_reports()
+    if ttnn.CONFIG.delete_reports_on_start:
+        ttnn.database.delete_reports()
     yield
     ttnn.tracer.disable_tracing()
